Changed unqualified use of Float to Float64, in preparation for adding new Float template.

+++ algebra/dense_differential.cc
@@ -27,6 +27,6 @@
 #include "algebra/dense_differential.h"
 
 namespace Ariadne {
-template class DenseDifferential<Float>;
-//template class Vector< DenseDifferential<Float> >;
+template class DenseDifferential<Float64>;
+//template class Vector< DenseDifferential<Float64> >;
 }
